[875] bff script; shelljs & isolate
