+++ package.json
@@ -51,6 +51,7 @@
     "react-typography": "^0.16.19",
     "s3-deploy": "^1.4.0",
     "sanitize-html": "^1.24.0",
+    "shelljs": "^0.8.4",
     "slash": "^3.0.0",
     "typography": "^0.16.19",
     "uuid": "^3.4.0"
